+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next
 ----
 
+2019.06.15.0
+------------
+
 * Add ``minidcos aws destroy`` and ``minidcos aws destroy-list`` commands.
 
 2019.06.10.0